Register controller when device comes up.
A controller was being registered with RFMonitor when ryu app was
initialized. Thus, it was not possible to keep a track on how many
devices a controller is connected to. Now the registration is done when
a device comes up. This helps in keeping exact count of devices to which
a controller is connected to.

Signed-off-by: Srijan Mishra <srijan23mishra@gmail.com>