Initialize controller address if not initialized.
The controller address is stored on CONF.ofp_listen_host config variable
provided by ryu API. If the parameter is not provided run time then the
controller is listens on localhost and the value of this variable is
uninitialized. So a check is made on init on variable and in case its
empty it is assigned value '127.0.0.1'.

Signed-off-by: Srijan Mishra <srijan23mishra@gmail.com>
Acked-by: Joe Stringer <joestringernz@gmail.com>